david9

+++ manuals.js
@@ -4,5 +4,5 @@
 
 SEK = 50433
 EUR = 5422
-const DONATION_TOTAL  = SEK + EUR*11;       // current total in SEK — update this each evening
+const DONATION_TOTAL  = int(SEK + EUR*11);       // current total in SEK — update this each evening
 const DONATION_GOAL   = 100000;  // your goal in SEK
